Add RestUser to documentation. Rework testing-other-systems docs. Rework gRPC example code (mainly moving the ClientInterceptor and abstract base User into its own file)
