+++ docs/testing-other-systems.rst
@@ -1,8 +1,8 @@
 .. _testing-other-systems:
 
-========================
-Testing non-HTTP systems
-========================
+===============================
+Testing other systems/protocols
+===============================
 
 Locust only comes with built-in support for HTTP/HTTPS but it can be extended to test almost any system. This is normally done by wrapping the protocol library and triggering a :py:attr:`request <locust.event.Events.request>` event after each call has completed, to let Locust know what happened.
 
@@ -14,10 +14,10 @@ Locust only comes with built-in support for HTTP/HTTPS but it can be extended to
 
     Some C libraries allow for other workarounds. For example, if you want to use psycopg2 to performance test PostgreSQL, you can use `psycogreen <https://github.com/psycopg/psycogreen/>`_. If you are willing to get your hands dirty, you may also be able to patch a library yourself, but that is beyond the scope of this documentation.
 
-Example: writing an XML-RPC User/client
-=======================================
+Example: XML-RPC
+================
 
-Lets assume we had an XML-RPC server that we wanted to load test.
+Lets assume we have an XML-RPC server that we want to load test.
 
 .. literalinclude:: ../examples/custom_xmlrpc_client/server.py
 
@@ -25,27 +25,63 @@ We can build a generic XML-RPC client, by wrapping :py:class:`xmlrpc.client.Serv
 
 .. literalinclude:: ../examples/custom_xmlrpc_client/xmlrpc_locustfile.py
 
-Example: writing a gRPC User/client
-=======================================
+Example: gRPC
+=============
 
-If you have understood the XML-RPC example, you can easily build a `gRPC <https://github.com/grpc/grpc>`_ User.
+We can also build a `gRPC <https://github.com/grpc/grpc>`_ User.
 
-The only significant difference is that you need to make gRPC gevent-compatible, by executing this code before opening the channel:
-
-.. code-block:: python
-
-    import grpc.experimental.gevent as grpc_gevent
-
-    grpc_gevent.init_gevent()
-
-Dummy server to test:
+Lets assume we have a gRPC server that we want to load test:
 
 .. literalinclude:: ../examples/grpc/hello_server.py
 
-gRPC client, base GrpcUser, interceptor for sending events to locust and example usage:
+The generic gRPC User base class sends events to Locust using an `interceptor <https://pypi.org/project/grpc-interceptor/>`_:
+
+.. literalinclude:: ../examples/grpc/grpc_user.py
+
+And an example test would look like this:
 
 .. literalinclude:: ../examples/grpc/locustfile.py
 
-As base class for interceptor is used `grpc-interceptor <https://pypi.org/project/grpc-interceptor/>` library.
+.. _testing-request-sdks:
 
-For more examples of user types, see `locust-plugins <https://github.com/SvenskaSpel/locust-plugins#users>`_ (it has users for WebSocket/SocketIO, Kafka, Selenium/WebDriver and more).
+requests-based libraries/SDKs
+=============================
+
+If you want to use a library that uses a `requests.Session <https://requests.readthedocs.io/en/latest/api/#requests.Session>`_ object under the hood you will most likely be able to skip all the above complexity.
+
+Some libraries allow you to pass a Session explicitly, like for example the SOAP client provided by `Zeep <https://docs.python-zeep.org/en/master/transport.html#tls-verification>`_. In that case, just pass it your ``HttpUser``'s :py:attr:`client <locust.HttpUser.client>`, and any requests made using the library will be logged in Locust.
+
+Even if your library doesn't expose that in its interface, you may be able to get it working by overwriting some internally used Session. Here's an example of how to do that for the `Archivist <https://github.com/jitsuin-inc/archivist-python>`_ client.
+
+.. literalinclude:: ../examples/sdk_session_patching/session_patch_locustfile.py
+
+
+Example: REST
+=============
+
+While the base HttpUser/FastHttpUser is capable of testing RESTful endpoints, it can be simplified by using a specialized subclass :py:class:`RestUser <locust.contrib.rest.RestUser>`. It extends FastHttpUser by adding the ``rest``-method, a wrapper around ``self.client.request()`` that:
+    
+* automatically passes ``catch_response=True``
+* automatically sets ``Content-Type`` and ``Accept`` headers to ``application/json`` (unless you have provided your own headers)
+* automatically checks that the response is valid json, parses it into an dict and saves it in a field called ``js`` in the response object.
+* catches any exceptions thrown in your ``with``-block and fails the sample (instead of crashing the task)
+
+.. code-block:: python
+
+    from locust.contrib.rest import RestUser
+    from locust import task
+
+    class MyUser(RestUser):
+        @task
+        def t(self):
+            with self.rest("POST", "/", json={"foo": 1}) as resp:
+                if resp.js and resp.js["bar"] != 1:
+                    resp.failure(f"Unexpected value of foo in response {resp.text}")
+
+For a complete example, see `resp_ex.py <https://github.com/locustio/locust/tree/master/examples/resp_ex.py>`_. That also shows how you can subclass :py:class:`RestUser <locust.contrib.rest.RestUser>` to provide behaviours specific to your API, like like always sending common headers or always applying some validation to the response.
+
+
+.. note::
+
+    For more examples of user types, see `locust-plugins <https://github.com/SvenskaSpel/locust-plugins#users>`_ (it has users for WebSocket/SocketIO, Kafka, Selenium/WebDriver and more).
+
